chore: Decouple Resynchronize class
