+++ mobile/src/modules/scanning/helpers/rescan.ts
@@ -7,8 +7,8 @@ import { fileNodes, invalidTracks, tracks } from "@/db/schema";
 
 import i18next from "@/modules/i18n";
 import { getTracks } from "@/api/track";
-import { Resynchronize } from "@/modules/media/services/Music";
 import { RecentList } from "@/modules/media/services/RecentList";
+import { Resynchronize } from "@/modules/media/services/Resynchronize";
 
 import { ToastOptions } from "@/lib/toast";
 import { batch } from "@/utils/promise";
